requirements.txt
numpy
scipy
pandas
matplotlib
pyyaml
tqdm
einops
gpytorch
gymnasium
rich
